depricated search files in BE, reorganized and combined functions in Workouts

+++ client/containers/Workouts.tsx
@@ -1,7 +1,6 @@
 import React, { useEffect, useState } from 'react';
 import Calendar from 'react-calendar';
 import 'react-calendar/dist/Calendar.css';
-// import WorkoutModal from '../modals/WorkoutModal';
 import { useNavigate, useLocation } from 'react-router-dom';
 import {
   updateSet,
@@ -9,60 +8,55 @@ import {
   deleteSet,
   postWorkout,
 } from '../api/workoutData';
-// import SetModal from '../modals/SetModal';
 import { useWorkoutStore, useAuthStore } from '../zustand';
 import ConfirmationModal from '../modals/ConfirmationModal';
 import EditForm from '../modals/EditForm';
-// import { postWorkout } from '../api/workoutData';
+import { WorkoutSet } from '../../types/types';
 
+// types for date, extended to array (range) if neccessary
 type ValuePiece = Date | null;
 type Value = ValuePiece | [ValuePiece, ValuePiece];
-// type date = string | null;
 
 function Workouts() {
   const [selectedDate, setSelectedDate] = useState(new Date());
-  const [value, onChange] = useState<Value>(new Date());
-  const [openWorkout, setOpenWorkout] = useState(false);
+  const [value, setValue] = useState<Value>(new Date());
   const [openCalendar, setOpenCalendar] = useState(false);
-  // const [workouts, setWorkouts] = useState<{ [exercise: string]: Array<{ reps: number; weight: number; exerciseId: number; setId: number }> }>({});
   const [selectedExercise, setSelectedExercise] = useState('');
-  // const [openSetModal, setOpenSetModal] = useState(false);
-  // for editing weight + reps for a set
-  const [editingSetId, setEditingSetId] = useState(null);
+  // state for editing weight + reps for a set
+  const [editingSetId, setEditingSetId] = useState<number | null>(null);
   const [editFormData, setEditFormData] = useState({ reps: '', weight: '' });
 
   const [isConfirmationModalOpen, setIsConfirmationModalOpen] = useState(false);
-  const [workoutToDelete, setWorkoutToDelete] = useState(null);
+  const [workoutToDelete, setWorkoutToDelete] = useState<number | null>(null);
 
   const { workouts, refreshWorkouts } = useWorkoutStore();
   const { token } = useAuthStore();
   const navigate = useNavigate();
   const location = useLocation();
-  // console.log('workouts: ', workouts)
 
-  // workout modal functions => might rework
-  // const openWorkoutModal = () => {
-  //   setOpenWorkout(true);
-  // };
-
-  // const closeWorkoutModal = () => {
-  //   setOpenWorkout(false);
-  // };
-  // Calendar modal functions
-  const openCalendarModal = () => {
-    setOpenCalendar(true);
-  };
-  const closeCalendarModal = (date: Date) => {
-    setSelectedDate(date);
-    setOpenCalendar(false);
-  };
-  const cancelCalendarModal = () => {
-    setOpenCalendar(false);
+  const currentDate = (date: Date) => {
+    // forms date string for display
+    const todaysDate = `${date.getMonth() + 1}/${date.getDate()}`;
+    return todaysDate;
   };
 
+  // combined into a single function with action for specification
+  const toggleCalendarModal = (date?: Date | null, action = '') => {
+    if (action === 'open') {
+      setOpenCalendar(true);
+    } else if (action === 'close' && date) {
+      setSelectedDate(date);
+      setOpenCalendar(false);
+    } else {
+      setOpenCalendar(false);
+    }
+  };
+
+  // put into another file?
   const handleAddSet = async (exercise: string) => {
     try {
       const unixtime = selectedDate.getTime();
+      // has to be here because workouts is from zustand, accessed by a hook
       const workoutData = {
         weight: Number(
           workouts[exercise][workouts[exercise].length - 1].weight,
@@ -81,10 +75,14 @@ function Workouts() {
     }
   };
 
-  const currentDate = (date: Date) => {
-    // forms date string for display
-    const todaysDate = `${date.getMonth() + 1}/${date.getDate()}`;
-    return todaysDate;
+  const handleDeleteSet = async (setId: number) => {
+    const unixtime = selectedDate.getTime();
+    const deleted = await deleteSet(setId, token);
+    if (deleted) {
+      refreshWorkouts(unixtime, token);
+    } else {
+      console.log('error deleting a set');
+    }
   };
 
   const handleExerciseClick = (exercise: string) => {
@@ -95,58 +93,47 @@ function Workouts() {
     }
   };
 
-  const handleDeleteSet = async (setId: number) => {
-    const unixtime = selectedDate.getTime();
-    // console.log(unixtime);
-    const deleted = await deleteSet(setId, token);
-    if (deleted) {
-      refreshWorkouts(unixtime, token);
-    }
-    // error case?
-  };
-
-  // fix type for workout parameter
-  const handleEditClick = (workout: any) => {
+  const handleEditClick = (workout: WorkoutSet) => {
     setEditingSetId(workout.setId);
-    setEditFormData({ reps: workout.reps, weight: workout.weight });
+    setEditFormData({
+      reps: String(workout.reps),
+      weight: String(workout.weight),
+    });
   };
 
   const handleSaveClick = async (setId: number) => {
-    // const updated =
     await updateSet(setId, editFormData, token);
     const unixtime = selectedDate.getTime();
-    // if (updated) {
     setEditingSetId(null);
     refreshWorkouts(unixtime, token);
-    // }
-    // error case?
-  };
-  // fix typing here
-  const openDeleteConfirmation = (exerciseId: any) => {
-    setWorkoutToDelete(exerciseId);
-    setIsConfirmationModalOpen(true);
-  };
-
-  const handleConfirmDelete = async () => {
-    if (workoutToDelete) {
-      const deleted = await deleteWorkout(workoutToDelete, token);
-      setWorkoutToDelete(null);
-      if (deleted) {
-        const unixtime = selectedDate.getTime();
-        refreshWorkouts(unixtime, token);
-      }
-    }
-    setIsConfirmationModalOpen(false);
-  };
-
-  // perhaps not needed, could combine with confirm function
-  const handleCancelDelete = () => {
-    setIsConfirmationModalOpen(false);
   };
 
   const handleSearchExerise = () => {
-    // should only pass serializable data, not functions
+    // navigate should only pass serializable data, not functions
     navigate('/search', { state: { selectedDate } });
+  };
+
+  // combined confirm function with action variable
+  const handleDeleteWorkout = async (
+    action: string,
+    exerciseId: number | null,
+  ) => {
+    setIsConfirmationModalOpen(false);
+    if (action === 'open') {
+      setWorkoutToDelete(exerciseId);
+      setIsConfirmationModalOpen(true);
+    } else if (action === 'confirm') {
+      if (workoutToDelete) {
+        const deleted = await deleteWorkout(workoutToDelete, token);
+        setWorkoutToDelete(null);
+        if (deleted) {
+          const unixtime = selectedDate.getTime();
+          refreshWorkouts(unixtime, token);
+        }
+      }
+    } else if (action === 'cancel') {
+      setIsConfirmationModalOpen(false);
+    }
   };
 
   // for fetching workout data
@@ -162,41 +149,32 @@ function Workouts() {
     fetchData();
 
     if (location.state && 'newSetId' in location.state) {
-      console.log('test:', location.state.exerciseName);
       setEditingSetId(location.state.newSetId);
       handleExerciseClick(location.state.exerciseName);
     }
-  }, [selectedDate, refreshWorkouts, location.state]);
+  }, [selectedDate, location.state]);
 
   return (
     <div className="flex flex-col relative h-screen">
       <button
         className="bg-gray-500 text-white font-bold py-2 px-4 rounded text-[30px] sticky top-0 z-10"
         type="button"
-        onClick={openCalendarModal}
+        onClick={() => toggleCalendarModal(undefined, 'open')}
       >
         <div className="hover:text-gray-400">{currentDate(selectedDate)}</div>
       </button>
       <div className="pt-3">
         {Object.keys(workouts).map((exercise: string) => (
-          <div className="pl-4" key={exercise}>
+          <div className="pl-4 pb-5" key={exercise}>
             <div className="flex justify-start">
               <div className="flex flex-col justify-start">
                 <button
                   type="button"
-                  onClick={() => {
-                    handleAddSet(exercise);
-                  }}
-                  className="px-5"
-                >
-                  <i className="material-icons text-[30px] text-red-500 hover:bg-slate-900 rounded">
-                    add
-                  </i>
-                </button>
-                <button
-                  type="button"
                   onClick={() =>
-                    openDeleteConfirmation(workouts[exercise][0].exerciseId)
+                    handleDeleteWorkout(
+                      'open',
+                      workouts[exercise][0].exerciseId,
+                    )
                   }
                   className="px-5"
                 >
@@ -216,7 +194,18 @@ function Workouts() {
               </button>
             </div>
             {selectedExercise === exercise && (
-              <div className="border-1 border-red-500">
+              <div className="">
+                <button
+                  type="button"
+                  onClick={() => {
+                    handleAddSet(exercise);
+                  }}
+                  className="px-5"
+                >
+                  <i className="material-icons text-[30px] text-red-500 hover:bg-slate-900 rounded">
+                    add
+                  </i>
+                </button>
                 <div className="flex justify-start px-2">
                   <span className="w-24 h-7 text-[15px] text-red-500 px-6">
                     Reps
@@ -269,8 +258,8 @@ function Workouts() {
         ))}
         <ConfirmationModal
           isOpen={isConfirmationModalOpen}
-          onConfirm={handleConfirmDelete}
-          onCancel={handleCancelDelete}
+          onConfirm={() => handleDeleteWorkout('confirm', null)}
+          onCancel={() => handleDeleteWorkout('cancel', null)}
           message="Are you sure you want to delete this workout?"
         />
       </div>
@@ -286,25 +275,24 @@ function Workouts() {
           <div className="pb-2">
             <button
               type="button"
-              onClick={() => cancelCalendarModal()}
+              onClick={() => toggleCalendarModal()}
               className="bg-red-500 hover:bg-white text-black hover:text-red-500 font-bold p-2 rounded-full inline-flex items-center justify-center w-10 h-10"
             >
               <span className="text-xl font-bold">X</span>
             </button>
           </div>
           <Calendar
-            onChange={onChange}
-            defaultValue={selectedDate}
-            onClickDay={closeCalendarModal}
+            onChange={setValue}
+            value={selectedDate}
+            onClickDay={(date) => toggleCalendarModal(date, 'close')}
+            tileClassName={({ date, view }) =>
+              view === 'month' && date.getTime() === selectedDate.getTime()
+                ? 'bg-red-500'
+                : null
+            }
           />
         </div>
       )}
-      {/* {openWorkout && (
-        <WorkoutModal
-          closeWorkoutModal={closeWorkoutModal}
-          selectedDate={selectedDate}
-        />
-      )} */}
     </div>
   );
 }
